requirements.txt
numpy
Keras
Pillow
tensorflow
PyYAML